Fix undefined error on first load
The call to fetch the widgets definition will initially return
`undefined`, but the code assumed it would be an object and would crash
on this. This in turn caused the component to error out and the `useSMR`
hook to never get a chance to actually fetch the data, so on a reload
the same issue would reoccur.